[enhancement] migrated DirichletSampler to Apache Commons RNG from custom implementation. [refactor] update dependencies and improve code structure - moved Element class to model package. [docs] added high-level architecture diagram to documentation assets.
