+++ src/main/java/com/medals/libsdatagenerator/util/CommonUtils.java
@@ -1,7 +1,7 @@
 package com.medals.libsdatagenerator.util;
 
 import com.medals.libsdatagenerator.controller.LIBSDataGenConstants;
-import com.medals.libsdatagenerator.service.Element;
+import com.medals.libsdatagenerator.model.Element;
 import org.apache.commons.cli.*;
 
 import java.io.File;
