+++ src/test/java/com/medals/libsdatagenerator/util/ElementPropertiesTest.java
@@ -1,10 +1,9 @@
 package com.medals.libsdatagenerator.util;
 
-import com.medals.libsdatagenerator.service.Element;
+import com.medals.libsdatagenerator.model.Element;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
